# SceneReader : Remove scene:path from context