Hashing the fileName plug for every scene location can cause significant
performance loss, in particular in situations where we are dynamically setting
the fileName using an expression.
This has the downside of preventing us from changing the fileName per
scene location, but giving that up seems like worth the advantage gained in
performance, given that we don't actually have a practical use for such feature.

---
@johnhaddon, I made this pull request based on your suggestion, and indeed it had a noticeable impact in performance in the dynamic scene builder.
However, given my lack of experience in the details of the code I changed, feel free to close the PR and make a new one yourself if you feel that patching the PR is more work :smile: