@ivanimanishi 2016-05-30
For the record, here are some performance stats:
Before:
```
  SceneBuilder.BuilderBox.GeoReader.SceneReader.out.object         30.667137 seconds
  SceneBuilder.BuilderBox.GeoReader.SceneReader.out.attributes     29.795906 seconds
  SceneBuilder.BuilderBox.GeoReader.SceneReader.out.transform      14.814098 seconds
  SceneBuilder.BuilderBox.GeoReader.SceneReader.out.bound          5.394537 seconds
```
After:
```
  SceneBuilder.BuilderBox.GeoReader.SceneReader.out.attributes     11.875723 seconds
  SceneBuilder.BuilderBox.GeoReader.SceneReader.out.transform      9.957635 seconds
  SceneBuilder.BuilderBox.GeoReader.SceneReader.out.object         8.343933 seconds
  SceneBuilder.BuilderBox.GeoReader.SceneReader.out.bound          2.368463 seconds
```