@ivanimanishi 2016-05-31
Tests passed now.