@johnhaddon 2016-05-31
> This has the downside of preventing us from changing the fileName per
scene location

I don't see that as a downside at all - changing the file name per scene location would allow totally invalid scenes to be created, so I think we can even call this a bug fix. Thanks for taking this one on - those speedups are rather pleasing!

So now, the big question, with this optimisation, what is the speed difference between the fancy SceneLoop scene building process and a more conventional static version?