@ivanimanishi 2016-05-31
@johnhaddon, I've updated the PR with your suggestions.
The name of the utility function is not exactly the one you suggested (I used `hashFileName`, since it's a bit more convenient to update the hash in-place, and it does conform with some of the other methods defined in the class), but if you think I should write a `fileNameHash` function that returns the hash to be added, I can do that.

> So now, the big question, with this optimisation, what is the speed difference between the fancy SceneLoop scene building process and a more conventional static version?

I don't have a comparison between the two at this time. But I have been able to do renders with both methods in a BRP and the total render time for each frame was very similar. But that scene wasn't particularly heavy, so it may not be a good test.

I definitely think we should at least update the Set to accept multiple names, so I can get rid of the "Tag" loop, as that is being called a lot of times at the moment.