@johnhaddon 2016-06-01
> @johnhaddon, I've updated the PR with your suggestions.
> The name of the utility function is not exactly the one you suggested.

Thanks - looks good. I've made one more inline comment where I think we can improve performance a touch, since that's our big focus here.

> I don't have a comparison between the two at this time. But I have been able to do renders with both methods in a BRP and the total render time for each frame was very similar. But that scene wasn't particularly heavy, so it may not be a good test.

My specific concern is that the SceneLoop approach will scale poorly as the number of assets (trips through the loop) is increased, rather than for single large assets. I think it's essential that we test this early on so we know what we're facing. I think there's potential for mitigating problems with a custom node designed for this exact purpose, so an early heads up on potential problems would be good.