@ivanimanishi 2016-06-01
Updated the PR to address @johnhaddon 's last comment.