@ivanimanishi 2016-06-01
The stats I posted earlier should be disregarded. I cannot reproduce the same values anymore for the initial state.
It seems that those results are likely caused by some temporary filesystem/caching issue.

Here is a more realistic effect of the changes in this PR:
```
Top plugs by number of hash processes :
before:
  SceneBuilder.BuilderBox.MainExpression.__execute                 224562
  BundleContents.Expression.__out.p0                               224562
after
  SceneBuilder.BuilderBox.MainExpression.__execute                 153231
  BundleContents.Expression.__out.p0                               153231

Top plugs by time spent in hash processes :
before
  EntityReader.Expression.__execute                                1.356765 seconds
after
  EntityReader.Expression.__execute                                0.920990 seconds
```