@johnhaddon 2016-06-01
Hmm - does that reduction in expression hashing result in an overall reduction in runtime, or is it offset by the added cost of the new context manipulation I wonder? I'm a little bit surprised that the reduction in the number of hash processes isn't bigger - those numbers imply that either the hash cache isn't especially effective in this case, or we're somehow still generating a lot of unique contexts to run the expression in.