@ivanimanishi 2016-06-02
So, I've ran a few more tests here, and it seems like even though we reduce the number of hash calls, the increased cost of creating new contexts without those variables seem to offset the gains, or even reverse them.

The differences are usually small either way.

There is also a small memory usage reduction with the changes in this PR, but they are not very significant (20M in a total around 3.6G).

So, I'm not sure if merging this PR is worth it. From the data I've been able to gather, no.