@johnhaddon 2016-06-03
I've chased similar optimisations in the past and had similarly disappointing results. I suspect that all isn't lost in the longer term though, because I think we can still make big improvements in the efficiency of Context creation. I'll close this for now, but if you could keep the branch around I think this might well be worth revisiting in the future. I'll make a note on the Context optimisation ticket to remind us to come back here.